requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.2
beautifulsoup4==4.12.2
aiohttp==3.9.1
python-dotenv==1.0.0